src/test/test_barcode.py
import os
from unittest.mock import patch, MagicMock
import pytest
# Importing the functions from the module
from re_barcode import capture_image, read_barcode, main

@patch('re_barcode.MockPiCamera')
@patch('re_barcode.sleep', return_value=None)
def test_capture_image(mock_sleep, MockPiCamera):
    file_path = 'test.jpg'
    capture_image(file_path)
    MockPiCamera.assert_called_once()
    mock_sleep.assert_called_once_with(2)

@patch('re_barcode.MockImage.open')
@patch('re_barcode.mock_decode')
def test_read_barcode(mock_decode, MockImage_open):
    file_path = 'test.jpg'
    MockImage_open.return_value = MagicMock()
    mock_decode.return_value = [MagicMock(data=b'1234567890', type='mock_type')]

    barcode_data = read_barcode(file_path)

    MockImage_open.assert_called_once_with(file_path)
    mock_decode.assert_called_once()
    assert barcode_data == '1234567890'

@patch('re_barcode.capture_image')
@patch('re_barcode.read_barcode', return_value='1234567890')
def test_main(mock_read_barcode, mock_capture_image):
    with patch('os.path.join', return_value='barcode.jpg'):
        main()
    mock_capture_image.assert_called_once()
    mock_read_barcode.assert_called_once_with('barcode.jpg')

if __name__ == "__main__":
    pytest.main()
